# Update MobileNet v2 to use fused kernels.