--- a/mobilenet/package.json
+++ b/mobilenet/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tensorflow-models/mobilenet",
-  "version": "2.0.4",
+  "version": "2.1.4",
   "description": "Pretrained MobileNet in TensorFlow.js",
   "main": "dist/index.js",
   "unpkg": "dist/mobilenet.min.js",
@@ -13,12 +13,12 @@
     "url": "https://github.com/tensorflow/tfjs-models.git"
   },
   "peerDependencies": {
-    "@tensorflow/tfjs-core": "~1.2.1",
-    "@tensorflow/tfjs-converter": "~1.2.1"
+    "@tensorflow/tfjs-core": "~1.2.11",
+    "@tensorflow/tfjs-converter": "~1.2.11"
   },
   "devDependencies": {
-    "@tensorflow/tfjs-core": "~1.2.1",
-    "@tensorflow/tfjs-converter": "~1.2.1",
+    "@tensorflow/tfjs-core": "~1.2.11",
+    "@tensorflow/tfjs-converter": "~1.2.11",
     "@types/jasmine": "~2.5.53",
     "babel-core": "^6.26.0",
     "babel-plugin-transform-runtime": "~6.23.0",
@@ -35,7 +35,7 @@
   },
   "scripts": {
     "build": "rimraf dist && tsc",
-    "publish-local": "yarn build && rollup -c && yalc push",    
+    "publish-local": "yarn build && rollup -c && yalc push",
     "lint": "tslint -p . -t verbose",
     "test": "ts-node run_tests.ts"
   },