#### Changes

- Update tfjs dependency to 1.2.11
- Update tfhub URLs

#### Benchmarks

Benchmarks are calculated for MobileNet v2 100, average of 50 runs.

|| Master  | update_mobilenet |
|------------- | ------------- | ------------ |
| # of kernels | 178 | 107 |
|MBP F32| 24.5ms  | 19.8ms *(1.24x)*  |
|Pixel3 F32| 99.1ms  | 88.5ms *(1.12x)*  |
|MBP F16| 22.7ms | 20.6ms *(1.1x)* |
|Pixel3 F16| 88.1ms | 77.8ms *(1.13x)* |
|iPhoneX| 23.4ms  | 21.6ms *(1.1x)*  |


<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-models/306)
<!-- Reviewable:end -->